requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
kiwisolver==1.4.9
packaging==25.0
pillow==12.0.0
pygame-ce==2.5.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
